+++ simplescn/cmdcom.py
@@ -39,12 +39,17 @@ def cmdloop(ip, use_unix=False, forcehash=None):
 
 def _single(address, use_unix, argv):
     command = argv[1]
+    body = {}
     if len(argv) >= 2:
-        stuff = argv[2]
+        args = argv[1:]
     else:
-        stuff = input()
+        args = shlex.split(input())
+    for elem in args:
+        splitted = elem.split("=", 1)
+        if len(splitted) == 2:
+            body[splitted[0]] = splitted[1]
     headers = {"Content-Type": "application/json; charset=utf-8"}
-    ret = do_request_simple(address, "/client/{}".format(command), stuff, headers, use_unix=use_unix)
+    ret = do_request_simple(address, "/client/{}".format(command), body, headers, use_unix=use_unix)
     print(ret)